feat: Added Health Bar
