+++ Assets/Prefabs/Enemies/enemy02.prefab
@@ -201,8 +201,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5ff3863895a58084bbfb9f1fbd22176c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  playerHealth: {fileID: 0}
-  damage: 5
+  damage: 20
 --- !u!95 &2805907496160769376
 Animator:
   serializedVersion: 5
