+++ Assets/Scripts/Core/ProjectileScript.cs
@@ -2,6 +2,8 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+namespace AGD
+{
 public class ProjectileScript : MonoBehaviour
 {
 
@@ -39,4 +41,5 @@ public class ProjectileScript : MonoBehaviour
             Destroy(gameObject);
         }
     }
+}
 }
