+++ Assets/Scripts/UI/HealthBar.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace AGD
+{
+    public class HealthBar : MonoBehaviour
+    {
+        public Slider slider;
+        public Gradient gradient;
+        public void SetMaxHealth(int health){
+            slider.maxValue = health;
+            slider.value = health;
+        }
+
+        public void setHealth(int health){
+            slider.value = health;
+        }
+    }
+}
