+++ Assets/Sprites/Bar.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1381c2d1754e5b8c213f038b0ad40d794e6371257b7a0c82b63f6a404b1e1050
+size 2358
