+++ Assets/TextMesh Pro/Fonts/alagard.ttf
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:12f1e58f79d18d2ef49d923d801baade892769289bfcb9444769b6a4f7847007
+size 13544
