+++ Assets/TextMesh Pro/Resources/Fonts & Materials/alagard SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1549d4bf8d26ae846972b7d46fa1e4a9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
